Added example project
