+++ SwiftVideoBackground/Sources/SwiftVideoBackground.h
@@ -1,0 +1,17 @@
+//
+//  SwiftVideoBackground.h
+//  SwiftVideoBackground
+//
+//  Created by Wilson Ding on 9/26/16.
+//  Copyright © 2016 Wilson Ding. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+//! Project version number for SwiftVideoBackground.
+FOUNDATION_EXPORT double SwiftVideoBackgroundVersionNumber;
+
+//! Project version string for SwiftVideoBackground.
+FOUNDATION_EXPORT const unsigned char SwiftVideoBackgroundVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <SwiftVideoBackground/PublicHeader.h>
