+++ SwiftVideoBackgroundTests/SwiftVideoBackgroundTests.swift
@@ -5,7 +5,6 @@
 //  Created by Wilson Ding on 9/28/16.
 //  Copyright © 2016 Wilson Ding. All rights reserved.
 //
-
 import XCTest
 import SwiftVideoBackground
